moved registry helper as seperate class
